# DentalCloudBackend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: DentalCloud API
  version: 0.1.0
  description: |
    Contrato inicial compartido para las fases 1–3 de DentalCloud.
    Las fechas y horas se transportan en ISO-8601 y la zona de negocio es
    America/El_Salvador. Los errores usan siempre el esquema ApiError.
servers:
  - url: http://localhost:8080
    description: Desarrollo local
...
paths:
...
  /api/patients/me/aftercare:
    get:
      tags: [Clinical]
      summary: Lista instrucciones post-cita publicadas propias
      operationId: listOwnAftercare
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Instrucciones publicadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AftercareInstruction'
  /api/aftercare:
    post:
      tags: [Clinical]
      summary: Crea una instrucción post-cita
      operationId: createAftercareInstruction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [appointmentId, title, body]
              properties:
                appointmentId: { type: string, format: uuid }
                title: { type: string, maxLength: 180 }
                body: { type: string, maxLength: 10000 }
                priority: { type: string, enum: [LOW, NORMAL, HIGH] }
                publish: { type: boolean, default: false }
      responses:
        '201':
          description: Instrucción creada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AftercareInstruction' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/aftercare/{id}/publicar:
    patch:
      tags: [Clinical]
      summary: Publica una instrucción post-cita propia
      operationId: publishAftercareInstruction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Instrucción publicada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AftercareInstruction' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/patients/me/documents:
    get:
      tags: [Clinical]
      summary: Lista metadatos de documentos visibles al paciente
      operationId: listOwnDocuments
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Documentos sin binarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClinicalDocument'
  /api/documentos:
    post:
      tags: [Clinical]
      summary: Carga un documento clínico privado
      operationId: uploadClinicalDocument
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, documentType, title]
              properties:
                file: { type: string, format: binary }
                patientId: { type: string, format: uuid, nullable: true }
                appointmentId: { type: string, format: uuid, nullable: true }
                planId: { type: string, format: uuid, nullable: true }
                documentType: { type: string, maxLength: 32 }
                title: { type: string, maxLength: 180 }
                visibleToPatient: { type: boolean, default: false }
      responses:
        '200':
          description: Documento almacenado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClinicalDocument' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/documentos/{id}:
    get:
      tags: [Clinical]
      summary: Consulta los metadatos de un documento autorizado
      operationId: getClinicalDocument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Metadatos del documento
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClinicalDocument' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/documentos/{id}/contenido:
    get:
      tags: [Clinical]
      summary: Descarga contenido con URL firmada
      operationId: downloadClinicalDocument
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: expiresAt
          in: query
          required: true
          schema: { type: integer, format: int64 }
        - name: signature
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: Binario privado }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/citas/{appointmentId}/documentos:
    get:
      tags: [Clinical]
      summary: Lista documentos vinculados a una cita
      operationId: listAppointmentDocuments
      security:
        - bearerAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Documentos de la cita
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ClinicalDocument' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/citas/{appointmentId}/aftercare:
    get:
      tags: [Clinical]
      summary: Lista instrucciones de una cita
      operationId: listAppointmentAftercare
      security:
        - bearerAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Instrucciones post-cita
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AftercareInstruction' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/treatment-plans/{planId}/documentos:
    get:
      tags: [Clinical]
      summary: Lista documentos vinculados a un plan
      operationId: listPlanDocuments
      security:
        - bearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Documentos del plan
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ClinicalDocument' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/documentos/{id}/publicar:
    patch:
      tags: [Clinical]
      summary: Publica un documento para el paciente
      operationId: publishClinicalDocument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Documento publicado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClinicalDocument' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    AftercareInstruction:
      type: object
      required: [id, appointmentId, patientId, title, body, priority]
      properties:
        id: { type: string, format: uuid }
        appointmentId: { type: string, format: uuid }
        patientId: { type: string, format: uuid }
        title: { type: string }
        body: { type: string }
        priority:
          type: string
          enum: [LOW, NORMAL, HIGH]
        publishedAt: { type: string, format: date-time, nullable: true }
    ClinicalDocument:
      type: object
      required: [id, patientId, documentType, title, mimeType, sizeBytes, createdAt]
      properties:
        id: { type: string, format: uuid }
        patientId: { type: string, format: uuid }
        appointmentId: { type: string, format: uuid, nullable: true }
        planId: { type: string, format: uuid, nullable: true }
        documentType: { type: string }
        title: { type: string }
        mimeType: { type: string }
        sizeBytes: { type: integer, minimum: 1, format: int64 }
        visibleToPatient: { type: boolean }
        downloadUrl: { type: string }
        createdAt: { type: string, format: date-time }
...
    ApiError:
      type: object
      required: [status, code, message, fieldErrors, traceId, timestamp]
      properties:
        status: { type: integer, example: 409 }
        code: { type: string, example: APPOINTMENT_SLOT_TAKEN }
        message: { type: string, example: Ese horario ya no está disponible. }
        fieldErrors:
          type: object
          additionalProperties: { type: string }
        traceId: { type: string, example: 8f2e8db4-4d11-4c3e-a37f-8be9c8bb75c9 }
        timestamp: { type: string, format: date-time }
  responses:
...
    Forbidden:
      description: Rol o ownership insuficiente
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example:
            status: 403
            code: ACCESS_DENIED
            message: No tienes permisos para realizar esta acción.
            fieldErrors: {}
            traceId: 8f2e8db4-4d11-4c3e-a37f-8be9c8bb75c9
            timestamp: '2026-08-04T15:30:00Z'
    NotFound:
      description: Recurso inexistente o no visible para el actor
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ValidationError:
      description: Regla de validación incumplida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example:
            status: 422
            code: VALIDATION_ERROR
            message: Uno o más campos no son válidos.
            fieldErrors:
              reason: El motivo es requerido.
            traceId: 8f2e8db4-4d11-4c3e-a37f-8be9c8bb75c9
            timestamp: '2026-08-04T15:30:00Z'
...